requirements.txt
# Function dependencies, for example:
# package>=version
requests==2.23.0
python-dotenv==0.12.0
